드디어.. 게시글 작성 됨요....

+++ src/main/java/com/springboot/board/domain/repository/AnswerRepository.java
@@ -1,9 +1,9 @@
 package com.springboot.board.domain.repository;
 
-import org.springframework.stereotype.Repository;
-import org.springframework.data.jpa.repository.JpaRepository;
 import com.springboot.board.domain.entity.Answer;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AnswerRepository extends JpaRepository<Answer, Integer> {
+public interface AnswerRepository extends JpaRepository<Answer, Long> {
 }
